% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\docType{data}
\name{codebook}
\alias{codebook}
\title{"codebook"}
\format{A data frame with 253 rows and 4 variables:
\describe{
\item{cityCode}{Sido Code}
\item{cityName}{Sido Name}
\item{sggCityCode}{sggCity Code}
\item{sggCityName}{sggCity Name}
}}
\source{
\url{http://info.nec.go.kr/}
}
\usage{
codebook
}
\description{
A data set containing the code by Sido, SGG, Precinct
}
\keyword{datasets}
