% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\docType{package}
\name{parliament}
\alias{parliament}
\alias{parliament-package}
\title{21th Parliament Election}
\description{
This package contains complete data for 21th Parliament Election.
2th Parliament Election formatted to be convenient for data analysis.
}
